feat(theme-builder): improve makeColorMix
* moved to a dedicated file
* allow color types other than Hct as input
* add function overloads
