+++ packages/theme-builder/src/index.ts
@@ -1,6 +1,7 @@
 export * from './core';
 export * from './dynamic-color';
-export * from './dynamic-color-data';
 export * from './dynamic-color-css';
+export * from './dynamic-color-data';
+export * from './dynamic-color-mix';
 export * from './dynamic-theme';
 export * from './from-image';
